refactor(db-helpers): rename `upserts into db-helpers
The new name describes the file contents better. There are some upsert helpers, but also some idemptent-insert ones.
